move logic to game scene
