+++ src/logic/ScoreManager.ts
@@ -1,5 +1,3 @@
-import { GameScene } from "../scenes/GameScene";
-
 export class ScoreManager {
     private score!: number;
     private highScore!: number;
