+++ src/objects/Shooter/Helpers/BulletCreator.ts
@@ -1,4 +1,4 @@
-import { logicGame } from "../../../game";
+// import { logicGame } from "../../../game";
 import { GameScene } from "../../../scenes/GameScene";
 import { ShootedBubble } from "../../ShootedBubble";
 import { Shooter } from "../Shooter";
@@ -18,7 +18,7 @@ export class BulletCreator {
     }
 
     private createShootedBubble() {
-        this.shooter.shootedBubble = new ShootedBubble(this.scene,28,28,logicGame.getColor());
+        this.shooter.shootedBubble = new ShootedBubble(this.scene,28,28,this.scene.colorManager.getTexture());
 
         Phaser.Display.Align.In.BottomCenter(this.shooter.shootedBubble,this.scene.mainZone, 0, -140);
         this.scene.add.existing(this.shooter.shootedBubble);
@@ -28,7 +28,7 @@ export class BulletCreator {
     }
 
     public createSecondBullet() {
-        this.shooter.secondBubllet = new ShootedBubble(this.scene,28,28,logicGame.getColor());
+        this.shooter.secondBubllet = new ShootedBubble(this.scene,28,28,this.scene.colorManager.getTexture());
 
         Phaser.Display.Align.In.BottomCenter(this.shooter.secondBubllet,this.scene.mainZone, 0, -140);
         Phaser.Display.Align.To.LeftCenter(this.shooter.secondBubllet,this.shooter.shootedBubble, -12, 20);
